third_party/WebKit/LayoutTests/push_messaging/push-messaging.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
description('Tests for the Push API.');

if (!window.testRunner)
    debug('This test can not run without testRunner');

window.jsTestIsAsync = true;

function testPushRegistrationError() {
    window.testRunner.setMockPushClientError('message');
    navigator.push.register('senderId').then(function() {
        testFailed('Success callback invoked unexpectedly.');
        testPushRegistrationSuccess();
    }, function(e) {
        error = e;
        shouldBeEqualToString('error.name', 'AbortError');
        testPushRegistrationSuccess();
    });
}
testPushRegistrationError();

function testPushRegistrationSuccess() {
    window.testRunner.setMockPushClientSuccess('endpoint', 'registrationId');
    navigator.push.register('senderId').then(function(reg) {
        registration = reg;
        shouldBeEqualToString('registration.pushEndpoint', 'endpoint');
        shouldBeEqualToString('registration.pushRegistrationId', 'registrationId');
        finishJSTest();
    }, function(e) {
        testFailed('Error callback invoked unexpectedly.');
        finishJSTest();
    });
}
</script>
</body>
</html>
