third_party/WebKit/LayoutTests/push_messaging/push-messaging-expected.txt

Tests for the Push API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.name is "AbortError"
PASS registration.pushEndpoint is "endpoint"
PASS registration.pushRegistrationId is "registrationId"
PASS successfullyParsed is true

TEST COMPLETE
